edit Email function moved
