+++ components/admin/EnrollmentsManagement.tsx
@@ -152,6 +152,9 @@ const EnrollmentList = ({
     startDate: "",
     endDate: null,
     availability: [{ day: "", startTime: "", endTime: "" }],
+    day: "",
+    startTime: "",
+    endTime: "",
     meetingId: "",
     paused: false,
     duration: 1,
