+++ lib/type-utils.ts
@@ -83,3 +83,4 @@ export const InterfaceToTableProfiles = (data: Profile) => {
   };
   return profile;
 };
+
